cpumask: fix sysfs topology bitmasks and lists
This patch fixes an issue introduced in commit 1fae5629. The issue
makes terminating zero bytes visible in sysfs topology files
containing bitmaps and lists. Example:

$ xxd /sys/devices/system/cpu/cpu0/topology/core_cpus
00000000: 3033 0a00                                03..
                 ^^
Note the terminating '\0' after the line feed.

The problem is that bitmap_print_to_buf() in lib/bitmap.c is used as
if it would return the number of characters actually printed to buf
(excluding terminating '\0'), but it returns number of bytes copied to
buf (including the '\0').

Fix by copying and returning only the characters in the string.